@@ -1,21 +1,22 @@
 import hre from "hardhat";
 import { expect } from "chai";
-import { BigNumber, Contract } from "ethers";
-import { SignerWithAddress } from "@nomiclabs/hardhat-ethers/signers";
-import { JsonRpcProvider } from "@ethersproject/providers";
+// import { parseEther } from "ethers";
+import { DiceGame, RiggedRoll, RiggedRoll__factory } from "../typechain-types";
+import { HardhatEthersSigner } from "@nomicfoundation/hardhat-ethers/signers";
 
 const { ethers } = hre;
 
 describe("🚩 Challenge 3: 🎲 Dice Game", function () {
-  let deployer: SignerWithAddress;
-  let diceGame: Contract;
-  let riggedRoll: Contract;
-  let provider: JsonRpcProvider;
+  let diceGame: DiceGame;
+  let riggedRoll: RiggedRoll;
+  let deployer: HardhatEthersSigner;
+  const { provider } = ethers;
 
-  const rollAmountString = ".002";
-  const rollAmount = ethers.utils.parseEther(rollAmountString);
+  const rollAmountString = "0.002";
+  const rollAmount = ethers.parseEther(rollAmountString);
 
   async function deployContracts() {
+    [deployer] = await ethers.getSigners();
     const DiceGame = await ethers.getContractFactory("DiceGame");
     diceGame = await DiceGame.deploy();
 
@@ -26,16 +27,15 @@ describe("🚩 Challenge 3: 🎲 Dice Game", function () {
       contractArtifact = "contracts/RiggedRoll.sol:RiggedRoll";
     }
 
-    const RiggedRoll = await ethers.getContractFactory(contractArtifact);
-    riggedRoll = await RiggedRoll.deploy(diceGame.address);
-
-    [deployer] = await ethers.getSigners();
-    provider = ethers.provider;
+    const diceGameAddress = await diceGame.getAddress();
+    const RiggedRoll = (await ethers.getContractFactory(contractArtifact)) as RiggedRoll__factory;
+    riggedRoll = await RiggedRoll.deploy(diceGameAddress);
   }
 
-  function fundRiggedContract() {
+  async function fundRiggedContract() {
+    const riggedRollAddress = await riggedRoll.getAddress();
     return deployer.sendTransaction({
-      to: riggedRoll.address,
+      to: riggedRollAddress,
       value: rollAmount,
     });
   }
@@ -45,17 +45,21 @@ describe("🚩 Challenge 3: 🎲 Dice Game", function () {
     while (true) {
       const latestBlockNumber = await provider.getBlockNumber();
       const block = await provider.getBlock(latestBlockNumber);
+      if (!block) {
+        return;
+      }
       const prevHash = block.hash;
       const nonce = await diceGame.nonce();
 
-      const hash = ethers.utils.solidityKeccak256(
+      const diceGameAddress = await diceGame.getAddress();
+      const hash = ethers.solidityPackedKeccak256(
         ["bytes32", "address", "uint256"],
-        [prevHash, diceGame.address, nonce],
+        [prevHash, diceGameAddress, nonce],
       );
 
-      const bigNum = BigNumber.from(hash);
-      expectedRoll = bigNum.mod(16);
-      if (expectedRoll.lte(5) == getRollLessThanFive) {
+      const bigInt = BigInt(hash);
+      expectedRoll = bigInt % 16n;
+      if (expectedRoll < 5n == getRollLessThanFive) {
         break;
       }
 
@@ -66,20 +70,30 @@ describe("🚩 Challenge 3: 🎲 Dice Game", function () {
   }
 
   describe("⚙  Setup contracts", function () {
+    console.log("asdf");
+
     it("Should deploy contracts", async function () {
+      console.log("1");
       await deployContracts();
-      expect(await riggedRoll.diceGame()).to.equal(diceGame.address);
+      console.log("2");
+
+      const diceGameAddress = await diceGame.getAddress();
+      console.log("3");
+
+      expect(await riggedRoll.diceGame()).to.equal(diceGameAddress);
     });
 
     it(`Should revert if balance is less than ${rollAmountString} ethers`, async function () {
+      console.log("dsssa");
       await expect(riggedRoll.riggedRoll()).to.be.reverted;
     });
 
     it("Should transfer sufficient eth to RiggedRoll", async function () {
       console.log("\t", "💸 Funding RiggedRoll contract");
       await fundRiggedContract();
-      const balance = await provider.getBalance(riggedRoll.address);
-      console.log("\t", "💲 RiggedRoll balance: ", ethers.utils.formatEther(balance));
+      const riggedRollAddress = await riggedRoll.getAddress();
+      const balance = await provider.getBalance(riggedRollAddress);
+      console.log("\t", "💲 RiggedRoll balance: ", ethers.formatEther(balance));
       expect(balance).to.gte(rollAmount, `Error when expecting DiceGame contract to have >= ${rollAmount} eth`);
     });
   });
@@ -88,17 +102,19 @@ describe("🚩 Challenge 3: 🎲 Dice Game", function () {
     it("Should call diceGame.rollTheDice for a roll <= 5", async () => {
       const getRollLessThanFive = true;
       const expectedRoll = await getRoll(getRollLessThanFive);
-      console.log("\t", "🎲 Expect roll to be less than or equal to 5. Dice Game Roll:", expectedRoll.toNumber());
+      console.log("\t", "🎲 Expect roll to be less than or equal to 5. Dice Game Roll:", Number(expectedRoll));
 
       const tx = await riggedRoll.riggedRoll();
-      await expect(tx).to.emit(diceGame, "Roll").withArgs(riggedRoll.address, rollAmount, expectedRoll);
+      const riggedRollAddress = await riggedRoll.getAddress();
+
+      await expect(tx).to.emit(diceGame, "Roll").withArgs(riggedRollAddress, rollAmount, expectedRoll);
       await expect(tx).to.emit(diceGame, "Winner");
     });
 
     it("Should not call diceGame.rollTheDice for a roll > 5", async () => {
       const getRollLessThanFive = false;
       const expectedRoll = await getRoll(getRollLessThanFive);
-      console.log("\t", "🎲 Expect roll to be greater than 5. Dice Game Roll:", expectedRoll.toNumber());
+      console.log("\t", "🎲 Expect roll to be greater than 5. Dice Game Roll:", Number(expectedRoll));
       console.log("\t", "◀  Expect riggedRoll to be reverted");
 
       await expect(riggedRoll.riggedRoll()).to.be.reverted;
@@ -108,15 +124,19 @@ describe("🚩 Challenge 3: 🎲 Dice Game", function () {
       console.log("\t", "💸 Funding RiggedRoll contract");
       await fundRiggedContract();
 
-      const prevBalance = await deployer.getBalance();
-      console.log("\t", "💲 Current RiggedRoll balance: ", ethers.utils.formatEther(prevBalance));
-      await riggedRoll.withdraw(deployer.address, provider.getBalance(riggedRoll.address));
+      const deployerPrevBalance = await provider.getBalance(deployer.address);
+      console.log("\t", "💲 Current RiggedRoll balance: ", ethers.formatEther(deployerPrevBalance));
+      const riggedRollAddress = await riggedRoll.getAddress();
+      const riggedRollBalance = await provider.getBalance(riggedRollAddress);
+      await riggedRoll.withdraw(deployer.address, riggedRollBalance);
 
-      const curBalance = await deployer.getBalance();
-      console.log("\t", "💲 New RiggedRoll balance: ", ethers.utils.formatEther(curBalance));
+      const deployerCurrentBalance = await provider.getBalance(deployer.address);
+      console.log("\t", "💲 New RiggedRoll balance: ", ethers.formatEther(deployerCurrentBalance));
 
-      expect(prevBalance.lt(curBalance), "Error when expecting RiggedRoll balance to increase when calling withdraw").to
-        .true;
+      expect(
+        deployerPrevBalance < deployerCurrentBalance,
+        "Error when expecting RiggedRoll balance to increase when calling withdraw",
+      ).to.true;
     });
   });
 });
